+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ### Next
 
+- Add `createPilingJsFromState` as a named export to simplify the initialization of a piling.js instance from an existing state.
 - Allow subsampling of previews
 - Allow initializing from previous state using `createPilingJs(element, { ... }, { initFromState: true })`
+- Return promise when running `importState()` that resolves once piling.js' state was overridden.
 
 _[Changes since v0.7.3](https://github.com/flekschas/piling.js/compare/v0.7.3...master)_